refactor: 스프린트내에 멤버로직 필요한 부분과 필요없는 log 제거
